+++ src/mobile.ts
@@ -20,6 +20,11 @@ $(function () {
   initHistory()
 
   const container = $('[pickupmobileapp-container]')
+  if (container.attr('data-demo') === '1') {
+    demoMode(container)
+    return
+  }
+
   let entrepotId = container.attr('data-entrepot-id')
   if (entrepotId === '') { entrepotId = undefined }
   const usePCat = container.attr('data-use-pcat') === '1'
@@ -88,3 +93,78 @@ $(function () {
 
   window.pickupMobileMachine = machine
 })
+
+/**
+ * This instanciate a false Machine, where we can easily test all states types.
+ * Use only for dev purpose.
+ * @param container
+ */
+function demoMode (container: JQuery): void {
+  const version = container.attr('data-modpickup-version') ?? '0'
+  const definition: {[key: string]: StateDefinition} = {}
+
+  definition.init = {
+    type: 'choice',
+    label: 'Accueil',
+    choices: [
+      {
+        label: 'Pick',
+        value: 'pick',
+        goto: 'pick'
+      },
+      {
+        label: 'Form',
+        value: 'form',
+        goto: 'form'
+      },
+      {
+        label: 'Show',
+        value: 'show',
+        goto: 'show'
+      },
+      {
+        label: 'Select',
+        value: 'select',
+        goto: 'select'
+      },
+      {
+        label: 'Unknown',
+        value: 'unknown',
+        goto: 'dontexist'
+      }
+    ]
+  }
+
+  definition.pick = {
+    type: 'pick',
+    label: 'Pick test',
+    key: 'demo',
+    primaryKey: 'rowid',
+    goto: 'unknown',
+    creationGoto: 'unknown',
+    creationLabel: 'New XXX',
+    fields: [
+      { name: 'field1', label: 'Demo 1', applyFilter: 'localeUpperCase' },
+      { name: 'field2', label: 'Demo 2' }
+    ]
+  }
+
+  definition.select = {
+    type: 'select',
+    label: 'Select test (not fully implemented)',
+    options: [
+      { label: 'Option 1', value: '1' },
+      { label: 'Option 2', value: '2' }
+    ]
+  }
+
+  const machine = new Machine(
+    'Demo',
+    version,
+    container.attr('data-user-id') ?? '',
+    definition
+  )
+  machine.init(container)
+
+  window.pickupMobileMachine = machine
+}